Cleaned up reset
Condensed to a single line.

Added the change to the puzzleSolved boolean, otherwise on subsequent
plays it thought the puzzle was already solved, leading to very
incorrect behavior.  During backtracking it would skip over cells
sometimes since they were solved.  It would eventually run back and
forth across solved and sort of solved cells and eventually have a
portion of the field "solved" but still be out of options, at which
point recursion would exit, with the puzzle unsolved.